Create README.md
It has been a while since I worked on a coding project.  Falldown Rebirth represents my foray back into coding.  I decided to program it in Python because it takes me back to the time when I first started programming, in QBASIC.  To me, Python is like a grown man's version of QBASIC, and Falldown is the perfect project to shake off the rust.

(Perhaps I should discuss more about the game... I'll take that as a TODO :-D)